feat:Add the galaxie and different colors and add transition from first page to the second
